\name{winners-package}
\alias{winners-package}
\alias{winners}
\docType{package}
\title{
\packageTitle{winners}
}
\description{
\packageDescription{winners}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{winners}
\packageIndices{winners}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{winners}

Maintainer: \packageMaintainer{winners}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
